Fix updated Hue US dimmer (RWL020) not having any device automation triggers
